argparser added

+++ pymailer.py
@@ -3,7 +3,6 @@ import argparse
 import base64
 import os
 import shutil
-import sys
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 
@@ -49,6 +48,9 @@ class colors:
     BOLD = '\033[1m'
     UNDERLINE = '\033[4m'
 
+# Global variable
+parser = {}
+
 
 def _get_credentials():
     store = oauth2client.file.Storage(CREDENTIAL_FILE_PATH)
@@ -56,12 +58,11 @@ def _get_credentials():
     if not credentials or credentials.invalid:
         flow = client.flow_from_clientsecrets(SECRET_FILE_PATH, SCOPES)
         flow.user_agent = APPLICATION_NAME
-        parser = argparse.ArgumentParser()
-        parser.add_argument("--noauth_local_webserver", default=True)
+        parser.add_argument("--noauth_local_webserver", action='store_true', default=True)
         parser.add_argument("--logging_level", default='ERROR')
         parser.add_argument("--auth_host_name", default='localhost')
         parser.add_argument("--auth_host_port", default=[8080, 8090])
-        flags = parser.parse_args()
+        flags = parser.parse_args(['--noauth_local_webserver'])
         credentials = tools.run_flow(flow, store, flags=flags)
         print '%sStoring credentials to %s%s' % (colors.BOLD, CREDENTIAL_FILE_PATH, colors.END)
     return credentials
@@ -131,39 +132,45 @@ def getopts(argv):
 
 
 if __name__ == '__main__':
-    myargs = getopts(sys.argv)
     email = ''
     subject = ''
     body = ''
-    try:
-        if '-e' in myargs:
-            email = myargs['-e']
-        else:
-            raise ValidationError("email id is required")
-
-        if '-s' in myargs:
-            subject = myargs['-s']
-        else:
-            raise ValidationError("mail subject is required")
-
-        if '-b' in myargs:
-            body = myargs['-b']
-        else:
-            raise ValidationError("mail body is required")
-    except Exception as e:
-        print "%sError: %s%s" % (colors.ERROR, str(e), colors.END)
-        exit()
-
-    if not os.path.exists(CREDENTIAL_FILE_PATH):
-        shutil.rmtree(CONF_PATH)
     if not os.path.exists(SECRET_FILE_PATH):
         print '%sNo secret file found, please check %s%s' % (colors.ERROR, GITHUB_LINK, colors.END)
         exit(-1)
-    if not os.path.exists(CONF_PATH):
-        print '%sIt seems you are running it first time%s' % (colors.BOLD, colors.END)
-        print '%sLet\'s just setup things%s\n\n' % (colors.BOLD, colors.END)
-        os.makedirs(CONF_PATH)
-        send(SENDER_EMAIL_ID, 'PyMailer :: setting up', 'Thank you for using PyMailer')
+    parser = argparse.ArgumentParser(prog='PyMailer')
+    parser.add_argument("-e", default='', help='to email')
+    parser.add_argument("-s", default='', help='subject')
+    parser.add_argument("-b", default='', help='body (in html)')
+    parser.add_argument("--setup", action='store_true', help='to setup')
+    flags = parser.parse_args()
+    if flags.setup or not os.path.exists(CONF_PATH):
+        if not os.path.exists(CREDENTIAL_FILE_PATH) and os.path.exists(CONF_PATH):
+            shutil.rmtree(CONF_PATH)
+        if not os.path.exists(CONF_PATH):
+            print '%sIt seems you are running it first time%s' % (colors.BOLD, colors.END)
+            print '%sLet\'s just setup things%s\n\n' % (colors.BOLD, colors.END)
+            os.makedirs(CONF_PATH)
+            send(SENDER_EMAIL_ID, 'PyMailer :: setting up', 'Thank you for using PyMailer')
+            exit()
+    else:
+        try:
+            if flags.e:
+                email = flags.e
+            else:
+                raise ValidationError("email id is required")
+            if flags.s:
+                subject = flags.s
+            else:
+                raise ValidationError("mail subject is required")
+            if flags.b:
+                body = flags.b
+            else:
+                raise ValidationError("mail body is required")
+        except Exception as e:
+            print "%sError: %s%s" % (colors.ERROR, str(e), colors.END)
+            parser.print_help()
+            exit()
 
     print "%smail-data: %s %s%s" % (colors.BLUE, email, subject, colors.END)
     send(email, subject, body)
